+++ dwara-db/src/data/sql/dwara_update_2_1_34_CP.sql
@@ -1,1 +1,1 @@
-on_hold
+INSERT INTO `action` (`id`, `type`) VALUES ('restore_tape_and_move_it_to_cp_server', 'complex');