+++ dwara-db/src/data/sql/dwara_update_2_1_34_Ingest.sql
@@ -1,1 +1,2 @@
 INSERT INTO `action` (`id`, `type`) VALUES ('restore_tape_and_move_it_to_cp_server', 'complex');
+INSERT INTO `action` (`id`, `type`) VALUES ('generate_mezzanine_proxies', 'complex');