+++ ReleaseNotes.md
@@ -10,9 +10,12 @@
 1) Apply the upgrade sql script(/dwara-db/src/data/sql/dwara_update_2_1_34_CP.sql)
 2) Ensure application.properties has the below entries
  
+overwrite the existing
 > restoreTmpLocationForVerification=/data/cpserver
+
+and create
 > restoreTmpLocationForBulkRestore=/mnt/cpserver
-> cpServerIp=172.18.1.15
+> cpServerIp=dummyIp
 
 # Dwara App Version - 2.1.43 (20 Sep 2022)